преобразовал блок отзывов в компонент
