+++ src/pages/Offer.tsx
@@ -5,9 +5,9 @@ import routes from '../router/routes';
 import { Offer as OfferData } from '../offer/types';
 import { Card } from '../offer/components/Card/Card';
 import { classNames } from '../utils/classNames';
-import { FeedbackForm } from '../offer/components/FeedbackForm';
 import { Rating } from '../rating/components/Rating';
 import offerRatingClassNames from '../offer/constants/offerRatingClassNames';
+import { Review } from '../reviews';
 
 export function Offer({ offers }: { offers: OfferData[] }) {
   const { id } = useParams<{ id: string }>();
@@ -156,44 +156,7 @@ export function Offer({ offers }: { offers: OfferData[] }) {
                   </p>
                 </div>
               </div>
-              <section className="offer__reviews reviews">
-                <h2 className="reviews__title">
-                  Reviews · <span className="reviews__amount">1</span>
-                </h2>
-                <ul className="reviews__list">
-                  <li className="reviews__item">
-                    <div className="reviews__user user">
-                      <div className="reviews__avatar-wrapper user__avatar-wrapper">
-                        <img
-                          className="reviews__avatar user__avatar"
-                          src="img/avatar-max.jpg"
-                          width={54}
-                          height={54}
-                          alt="Reviews avatar"
-                        />
-                      </div>
-                      <span className="reviews__user-name">Max</span>
-                    </div>
-                    <div className="reviews__info">
-                      <div className="reviews__rating rating">
-                        <div className="reviews__stars rating__stars">
-                          <span style={{ width: '80%' }} />
-                          <span className="visually-hidden">Rating</span>
-                        </div>
-                      </div>
-                      <p className="reviews__text">
-                        A quiet cozy and picturesque that hides behind a a river
-                        by the unique lightness of Amsterdam. The building is
-                        green and from 18th century.
-                      </p>
-                      <time className="reviews__time" dateTime="2019-04-24">
-                        April 2019
-                      </time>
-                    </div>
-                  </li>
-                </ul>
-                <FeedbackForm />
-              </section>
+              <Review className="offer__reviews" reviews={offer.reviews} />
             </div>
           </div>
           <section className="offer__map map" />
